Refine example Remo debug surfaces
