+++ examples/ios/RemoExamplePackage/Sources/RemoExampleFeature/ConfettiOverlay.swift
@@ -1,9 +1,5 @@
 import SwiftUI
 
-#if canImport(UIKit)
-import UIKit
-#endif
-
 struct ConfettiOverlay: View {
     @State private var particles: [ConfettiParticle] = []
 
@@ -16,17 +12,17 @@ struct ConfettiOverlay: View {
                     .position(particle.position)
                     .opacity(particle.opacity)
             }
+            .onAppear { startConfetti(in: geo.size) }
         }
         .allowsHitTesting(false)
         .ignoresSafeArea()
-        .onAppear { startConfetti() }
     }
 
-    private func startConfetti() {
+    private func startConfetti(in size: CGSize) {
         particles = (0..<80).map { _ in
             ConfettiParticle(
                 position: CGPoint(
-                    x: CGFloat.random(in: 0...UIScreen.main.bounds.width),
+                    x: CGFloat.random(in: 0...max(size.width, 1)),
                     y: -20
                 ),
                 color: [Color.red, .blue, .green, .yellow, .purple, .orange, .pink].randomElement()!,
@@ -37,7 +33,7 @@ struct ConfettiOverlay: View {
 
         for i in particles.indices {
             let delay = Double.random(in: 0...0.5)
-            let targetY = CGFloat.random(in: 200...UIScreen.main.bounds.height + 100)
+            let targetY = CGFloat.random(in: 200...max(size.height, 200) + 100)
             let targetX = particles[i].position.x + CGFloat.random(in: -80...80)
 
             withAnimation(.easeOut(duration: Double.random(in: 1.5...2.5)).delay(delay)) {
